back-end/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
click==8.1.8
fastapi==0.115.10
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
redis==5.2.1
setuptools==75.8.2
simplejson==3.20.1
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
python-multipart==0.0.20
qrcode
pillow